% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessHWSD.R
\name{accessHWSD}
\alias{accessHWSD}
\title{Download and load HWSD}
\usage{
accessHWSD(
  downloadFolder,
  HWSD_proj = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{downloadFolder}{locatin to download the files to}

\item{HWSD_proj}{string defining the crs projection}
}
\value{
a list contining the raster of the mapping units and a list of the data tables supporting this raster
}
\description{
This function downloads the HWSD to a target folder from the project webpage, and unzips it.
It then loads the map of the mapping units as a raster object and the MS Access Database as a list of tables.
You will need to install mdbtools on your system to use Hmisc::mdb.get
}
