% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSOC.R
\name{makeSOC}
\alias{makeSOC}
\title{Calculate soil organic carbon stocks}
\usage{
makeSOC(hwsd_raster, hwsd_data, verbose = FALSE)
}
\arguments{
\item{hwsd_raster}{a raster object we will ratify and return with annotated values. Be careful with the size here!}

\item{hwsd_data}{a list containing the dataframe `HWSD_DATA` that has the reference values}

\item{verbose}{a boolean that dumps messages on function status}
}
\value{
a ratified raster object with top and sub soil organic carbon stocks
}
\description{
This function calculates the organic soil carbon stocks from the data structure and returns it as a ratified raster.
Note that there are several options on how to calculate this soil carbon stock and we return both the coarse fraction corrected and uncorrected estimates using the `BULK.DENSITY` values.
Note that we did NOT use the `REF.BULK.DENSITY` calculations here and such a choice is debatable.
Use your best judgement after consulting the project documentation.
}
